[fix] Updated ConversationPresenter to work with pluck_map 1.0 {{err:33053}} (1h)
It was both passing in raw strings instead of wrapping them with `Arel.sql()` and it was trying to order based on a column alias which doesn't exist in the query generated by pluck_map 1.0 (it uses __pluck_map_0, __pluck_map_1, etc. instead and aliases after the query); since it's a Backbone view that will order on the client anyways, it's not a huge issue if we don't present them ordered in the JSON we send the client.